CRUD operations for Person

+++ src/main/java/com/mic3/personservice/repository/PersonRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Syed Wajid
  * JPA repository for person
  */
-public interface PersonRepository extends JpaRepository<Person, Integer> {
+public interface PersonRepository extends JpaRepository<Person, Long> {
 }
